moved Twitter Login routes for verification of pattern full usage
